✨ : #86 - 리마인드 time 숫자 vs string 변환하는 로직 각 페이지 및 컴포넌트에 추가
