+++ src/apis/client/postNewPlan.ts
@@ -1,18 +1,6 @@
 import { DOMAIN } from '@/constants/api';
+import { PostNewPlanRequestBody } from '@/types/apis/plan/PostNewPlan';
 import { axiosInstanceClient } from '../axiosInstanceClient';
-
-export interface PostNewPlanRequestBody {
-  title: string;
-  description: string;
-  remindTotalPeriod: number;
-  remindTerm: number;
-  remindDate: number;
-  remindTime: number;
-  isPublic: boolean;
-  tags: string[];
-  messages: string[];
-  icon: number;
-}
 
 export const postNewPlan = async (body: PostNewPlanRequestBody) => {
   const { data } = await axiosInstanceClient.post(DOMAIN.POST_PLANS, {
