+++ src/app/(header)/edit/[planId]/page.tsx
@@ -8,6 +8,8 @@ import { useEditPlanMutation } from '@/hooks/apis/useEditPlanMutation';
 import { useGetRemindQuery } from '@/hooks/apis/useGetRemindQuery';
 import { EditPlanData } from '@/types/apis/plan/EditPlan';
 import { RemindItemType, RemindOptionType } from '@/types/components/Remind';
+import { changeRemindTimeToNumber } from '@/utils/changeRemindTimeToNumber';
+import { changeRemindTimeToString } from '@/utils/changeRemindTimeToString';
 import { checkIsSeason } from '@/utils/checkIsSeason';
 import { decideRemindDate } from '@/utils/decideRemindDate';
 import classNames from 'classnames';
@@ -39,64 +41,6 @@ export default function EditPage({ params }: { params: { planId: string } }) {
     checkIsSeason(),
   );
 
-  // const remindData: RemindData = {
-  //   isRemindable: true,
-  //   remindTime: 9,
-  //   remindDate: 1,
-  //   remindTerm: 1,
-  //   remindTotalPeriod: 12,
-  //   remindMessageList: [
-  //     {
-  //       remindMonth: 3,
-  //       remindDate: 15,
-  //       remindMessage: '리마인드 받았지만 만료되서 피드백 0%로 처리 ',
-  //       isReminded: true,
-  //       isFeedback: false,
-  //       feedbackId: 12,
-  //       rate: 0,
-  //       isExpired: true,
-  //       endMonth: 12,
-  //       endDate: 1,
-  //     },
-  //     {
-  //       remindMonth: 6,
-  //       remindDate: 15,
-  //       remindMessage: '리마인드 받아서 피드백함',
-  //       isReminded: true,
-  //       isFeedback: true,
-  //       feedbackId: 12,
-  //       rate: 75,
-  //       isExpired: true,
-  //       endMonth: 12,
-  //       endDate: 1,
-  //     },
-  //     {
-  //       remindMonth: 9,
-  //       remindDate: 15,
-  //       remindMessage: '예시',
-  //       isReminded: true,
-  //       isFeedback: false,
-  //       feedbackId: 12,
-  //       rate: 0,
-  //       isExpired: false,
-  //       endMonth: 12,
-  //       endDate: 1,
-  //     },
-  //     {
-  //       remindMonth: 12,
-  //       remindDate: 15,
-  //       remindMessage: '예시',
-  //       isReminded: false,
-  //       isFeedback: false,
-  //       feedbackId: 12,
-  //       rate: 0,
-  //       isExpired: false,
-  //       endMonth: 12,
-  //       endDate: 1,
-  //     },
-  //   ],
-  // };
-
   // 2. 계획 수정을 위해 관리되어야 하는 state 및 핸들러의 기본값에 1번에서 받은 값을 넣어준다.
   const [title, setTitle] = useState(planData.title);
   const [description, setDescription] = useState(planData.description);
@@ -119,7 +63,7 @@ export default function EditPage({ params }: { params: { planId: string } }) {
     TotalPeriod: remindData.remindTotalPeriod,
     Term: remindData.remindTerm,
     Date: remindData.remindDate,
-    Time: remindData.remindTime,
+    Time: changeRemindTimeToNumber(remindData.remindTime),
   });
 
   const handleChangeRemindOption = (
@@ -212,7 +156,7 @@ export default function EditPage({ params }: { params: { planId: string } }) {
       remindTotalPeriod: remindOptions.TotalPeriod,
       remindTerm: remindOptions.Term,
       remindDate: remindOptions.Date,
-      remindTime: remindOptions.Time,
+      remindTime: changeRemindTimeToString(remindOptions.Time),
       isPublic: isPublic,
       canRemind: isRemindOn,
       canAjaja: canAjaja,
@@ -222,7 +166,7 @@ export default function EditPage({ params }: { params: { planId: string } }) {
       }),
     };
 
-    editPlanAPI({ planId: parseInt(planId, 10), planData: editPlanData }); // 계획 수정 API 호출
+    editPlanAPI({ planId: parseInt(planId, 10), planData: editPlanData });
 
     console.log(`${planId}에 해당하는 계획 수정 API 호출 `);
   };
