+++ src/types/components/Remind.ts
@@ -7,7 +7,7 @@ export type RemindOptionType = {
   TotalPeriod: number;
   Term: number;
   Date: number;
-  Time: number;
+  Time: 9 | 13 | 20;
 };
 
 export interface RemindItemType {
@@ -33,7 +33,7 @@ export interface ReadOnlyRemindItemData {
 
 export interface RemindData {
   isRemindable: boolean;
-  remindTime: number;
+  remindTime: 'Morning' | 'Afternoon' | 'Evening';
   remindDate: number;
   remindTerm: number;
   remindTotalPeriod: number;
